complete check for blocked move

+++ CheckMove.h
@@ -7,8 +7,10 @@ int CheckMove(char activePLayer, unsigned int startrow, unsigned int startcolumn
         {  
             // Pawn
             case 1:
+                // standart move
                 if(destcolumn == startcolumn)
                 {
+                    // move direction decided by activePlayer
                     if(activePLayer)
                     {
                         if((destrow == startrow - 1) || ((startrow == 6) && (destrow == 4)))
@@ -19,11 +21,13 @@ int CheckMove(char activePLayer, unsigned int startrow, unsigned int startcolumn
                         if((destrow == startrow + 1) || ((startrow == 1) && (destrow == 3)))
                             return 1;
                     }
-                } 
-                else if(field[destrow][destcolumn][0] && field[destrow][destcolumn][1] != activePLayer)
+                }
+                // taking a piece (diagonal)
+                else if(field[destrow][destcolumn][0] && field[destrow][destcolumn][1] != (unsigned int)activePLayer)
                 {
                     if((destcolumn - startcolumn == 1) || (startcolumn - destcolumn == 1))
                     {
+                        // move direction
                         if(activePLayer && (destrow == startrow - 1))
                             return 1;
                         else if(!activePLayer && (destrow == startrow + 1))
@@ -38,21 +42,25 @@ int CheckMove(char activePLayer, unsigned int startrow, unsigned int startcolumn
                 break;
             // Knight
             case 3:
+                // move : 1 field vertically or horizontally and then 2 the other of the two
                 if(((abs((int)destrow - (int)startrow) == 1) && (abs((int)destcolumn - (int)startcolumn) == 2)) || ((abs((int)destcolumn - (int)startcolumn) == 1) && (abs((int)destrow - (int)startrow) == 2)))
                     return 1;
                 break;
-            // Bishops
+            // Bishop
             case 4:
+                // diagonal := x - difference = y - diff.
                 if(abs((int)destrow - (int)startrow) == abs((int)destcolumn - (int)startcolumn))
                     return 1;
                 break;
             // Queen
             case 5:
+                // combination of rook and bishop
                 if((destrow == startrow) || (destcolumn == startcolumn) || abs((int)destrow - (int)startrow) == abs((int)destcolumn - (int)startcolumn))
                     return 1;
                 break;
             // King
             case 6:
+                // move : 1 in any direction
                 if((abs((int)destrow - (int)startrow) < 2) && (abs((int)destcolumn - (int)startcolumn) < 2))
                     return 1;
                 break;
